requirements.txt
selenium==4.20.0
PyYAML==6.0.1
flask==3.0.3
psutil==5.9.8
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
